Fix checkstyle issues and refactor FieldMap package
